Restructure PPAs according to new week structure

+++ ppa/ppa_6.1/evaluation_file.sh
@@ -6,7 +6,7 @@ err(){
 }
 
 req=( "mktemp" "diff" "basename" "col" "find" "pushd" "popd" )
-executable="swapcase.sh"
+executable="backup.sh"
 for i in "${req[@]}"; do
   command -v "$i" > /dev/null 2>&1 || err "$i is not installed"
 done
@@ -26,8 +26,26 @@ cat >script.sh <<EOF
 rand_dir=\$(mktemp -d XXXXXX)
 pushd "\$rand_dir" > /dev/null || exit 1
 
-cat > words.txt
-bash "../\$(dirname "\${BASH_SOURCE[0]}")/$executable" 2>&1 < /dev/null
+mkdir -p source destination >/dev/null || exit 2
+
+while read -r line; do
+  if [[ "\${line: -1}" = "/" ]]; then
+    mkdir -p "\$line" >/dev/null || exit 3
+  else
+    mkdir -p "\$(dirname "\$line")" >/dev/null || exit 4
+    touch "\$line" >/dev/null || exit 5
+  fi
+done
+
+exec="../\$(dirname "\${BASH_SOURCE[0]}")/$executable"
+
+[[ -r "\$exec" ]] || exit 1
+
+bash "\$exec" 2>&1 < /dev/null
+
+tree source | sed '/^$/q'
+echo "---"
+tree destination | sed '/^$/q'
 
 popd > /dev/null || exit 1
 [[ -d "\$rand_dir" ]] && rm "\${rand_dir?}" -rf
